requirements.txt
fastapi~=0.112.0
uvicorn~=0.30.5
pydantic
pydantic-settings
sentence-transformers
chromadb~=0.5.5
langchain~=0.2.12
openai
python-multipart
torch~=2.4.0
transformers~=4.44.0
aiohttp~=3.10.3
uvicorn
pymupdf
fitz~=0.0.1.dev2
python-docx
websockets~=12.0